Add files via upload
This Python program has a class called TemperatureConverter that changes temperatures between Celsius and Fahrenheit. It checks if the input is correct, makes sure the temperature isn't below absolute zero, and shows clear error messages if something is wrong. The program also shows examples of how to use the class.